04-blog-to-multiplatform/pipeline/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
httpx==0.28.1
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.19
python-ulid==3.0.0
python-dotenv==1.0.1
watchfiles==1.0.4
starlette==0.41.3
anthropic>=0.40.0
